Modify the docker-compose.yml and dockerfile format normalization
