+++ CITA_Monitor_Agent_Docker/DockerCompose_Files/citaMonitorAgent_exporter/Dockerfile
@@ -1,0 +1,12 @@
+## 一个基础的 python 运行环境
+FROM python:3
+## 设置工作目录
+WORKDIR /config
+## 将当前系统文件夹内容复制到容器的 app 目录
+ADD . /config
+## 安装必要的依赖包
+RUN pip3 install -r requirements.txt
+RUN mv cita-cli /bin/cita-cli && chmod +x /bin/cita-cli
+ENV OpenPort=1920
+## 运行命令
+CMD "python3" "cita_agent_by_cita-cli.py" "$Node" "$OpenPort"
