+++ CITA_Monitor_Agent_Docker/DockerCompose_Files/citaMonitorAgent_exporter/requirements.txt
@@ -1,0 +1,2 @@
+prometheus-client==0.5.0
+flask
