--- CITA_Monitor_Agent_Docker/DockerImage_Build_Files/requirements.txt
@@ -1,2 +1,0 @@
-prometheus-client==0.5.0
-flask
