--- cita_agent_by_docker/DockerFile_Build/requirements.txt
@@ -1,2 +1,0 @@
-prometheus-client==0.5.0
-flask
